% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcor.R
\name{cal_pcor}
\alias{cal_pcor}
\title{pltPcorrelation}
\usage{
cal_pcor(pcor_dt = pcor_dt, ...)
}
\arguments{
\item{pcor_dt}{A dataframe of calculate in Partial correlation.}
}
\value{
cal_pcor
}
\description{
pltPcorrelation
}
\author{
Qiusheng WU
}
