% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pltpcor.R
\name{pltpcor}
\alias{pltpcor}
\title{pltPcorrelation}
\usage{
pltpcor(obj = obj, low = "green", high = "red", mid = "blue", ...)
}
\value{
pltpcor
}
\description{
pltPcorrelation
}
\examples{
library(pltPcorrelation)
data(mydt)
res <- cal_pcor(pcor_dt =mydt)
pltpcor(res,high = "red",low = "blue",mid = "gray")
}
\author{
Qiusheng WU
}
